+++ src/components/themed/AddressQr.tsx
@@ -33,7 +33,7 @@ export const AddressQr = (props: Props) => {
 
   return (
     <View style={styles.container}>
-      <QrCode data={encodedUri} tokenId={tokenId} wallet={wallet} onPress={() => onPress(encodedUri)} marginRem={0} />
+      <QrCode data={encodedUri} tokenId={tokenId} pluginId={wallet.currencyInfo.pluginId} onPress={() => onPress(encodedUri)} marginRem={0} />
     </View>
   )
 }